client/inferences/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
albumentations==1.4.3
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
av==12.3.0
backcall==0.2.0
base==0.0.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.103
botocore==1.34.103
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.10.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
easyocr==1.7.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.3
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.6.0
gdown==5.1.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==2.10
ifaddr==0.2.0
ImageHash==4.3.1
imageio==2.34.0
imgviz==1.7.5
install==1.3.5
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.4.1
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.4
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
namex==0.0.8
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
onnx==1.16.0
onnxruntime==1.17.3
onnxscript==0.1.0.dev20240413
opencv-python==4.7.0.72
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==4.25.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyclipper==1.3.0.post5
pycparser==2.22
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
roboflow==1.1.26
rpds-py==0.19.1
s3transfer==0.10.1
sahi==0.11.15
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
supervision==0.20.0
sympy==1.12
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.3.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
ultralytics @ git+https://github.com/ultralytics/ultralytics.git@9d48190e6d624df45714b089491e304525326db1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
yolov5==7.0.13
yolov8==0.0.2
